feat: Phase 2 — LikesView top-level destination with search and pagination
