+++ docs/features/backend-shares/EXECUTION_LOG.md
@@ -1,0 +1,66 @@
+# Execution Log — backend-shares
+
+**Started**: 2026-04-22
+**Plan**: `./PLAN.md`
+**Executor**: backend-engineer agent (autonomous mode per Dom directive "Run through it all go")
+
+---
+
+## Baseline state discovered
+
+- `xomify-infrastructure` master at `6c8455f` (clean, modulo untracked `.DS_Store` + `.github/workflows/claude-issues.yml`).
+- `xomify-backend` master at `65f9fbe` (clean, modulo untracked `.claude/prompts/` + `.github/workflows/claude-issues.yml`).
+- Merged shares-and-salvage PR (`#65` backend, `#65` infra) already landed:
+  - `lambdas/shares_create/`, `shares_feed/`, `shares_react/`, `invites_create/`, `invites_accept/` exist with handler implementations (different schema than this sub-feature plan calls for).
+  - `lambdas/common/shares_dynamo.py`, `invites_dynamo.py`, `interactions_dynamo.py` exist with partial functions.
+  - Terraform `locals.shares_lambdas` has 3 entries (`create`, `feed`, `react`); invites has 2 (`create`, `accept`).
+- `lambdas/common/constants.py` already exports `SHARES_TABLE_NAME`, `INVITES_TABLE_NAME`, `SHARE_INTERACTIONS_TABLE_NAME`.
+- Baseline `pytest`: 26 passed (ignoring 2 modules that need aiohttp/cffi — unrelated, pre-existing).
+
+## Schema ambiguity vs plan
+
+Existing `shares_create` handler uses `{email, type: wrapped|release_radar|track|playlist, payload: dict, caption}`.
+Plan calls for `{email, trackId, trackUri, trackName, artistName, albumName, albumArtUrl, caption?, moodTag?, genreTags?}`.
+
+**Decision**: follow plan. The plan was written AFTER the merged shares-and-salvage work and is the Ready source of truth. Live DDB table is agnostic (PK `shareId`, free-form attributes under GSI `email-createdAt-index`), so schema change is handler-layer only. Flagging in report-back.
+
+## Phase A — Terraform PR
+
+- Branch `feat/backend-shares-tf` off `origin/master` (@ `6c8455f`). Created.
+- Edited `terraform/lambdas_shares.tf` — appended `delete` + `user` entries to `local.shares_lambdas`. Left the 3 existing entries and the `aws_lambda_function.shares` `for_each` block untouched.
+- `terraform fmt terraform/` — no diff (already formatted).
+- `terraform plan` — **skipped locally**. No `terraform.tfvars` / no AWS creds in shell. Required `client_id`, `client_secret`, `api_access_token`, `api_secret_key` variables have no defaults. CI will catch drift on PR. Expected diff per plan: 2 new `aws_lambda_function.shares["delete"|"user"]` resources + API Gateway method/integration/permission resources under `module.api`. Zero changes to tables, IAM, authorizer.
+- Commit `1a9be45` pushed; PR https://github.com/Xomware/xomify-infrastructure/pull/67 opened. Left unmerged per instructions (Dom sequences).
+
+## Phase B — Backend handlers PR
+
+- Branch `feat/backend-shares-handlers` off `origin/master` (@ `65f9fbe`). Created.
+- `lambdas/common/constants.py`: added `SHARES_EMAIL_INDEX` and `INVITE_URL_TEMPLATE` next to existing shares/invites table names. `SHARES_TABLE_NAME`, `SHARE_INTERACTIONS_TABLE_NAME`, `INVITES_TABLE_NAME` already present.
+- `tests/conftest.py`: added `SHARES_TABLE_NAME`, `SHARE_INTERACTIONS_TABLE_NAME`, `INVITES_TABLE_NAME`, `SHARES_EMAIL_INDEX`, `INVITE_URL_TEMPLATE` to `_TEST_ENV_VARS`.
+- `lambdas/common/shares_dynamo.py`: rewritten per plan. `create_share` now takes denormalized Spotify metadata (trackId/trackUri/trackName/artistName/albumName/albumArtUrl) plus optional caption/moodTag/genreTags. Added `delete_share`, paginated `list_shares_for_user` returning `(items, next_before)`, and `query_feed_for_emails` with `ThreadPoolExecutor(max_workers=10)`.
+- `lambdas/common/invites_dynamo.py`: rewritten per plan. `generate_invite_code` produces 8-char base32 uppercase. `create_invite` uses `attribute_not_exists(inviteCode)` for collision safety. `consume_invite` does atomic `UpdateItem` with `attribute_not_exists(consumedAt) AND expiresAt > :now`. `list_invites_by_sender` uses Scan+FilterExpression (v1 — no sender GSI). `count_outstanding_invites_for_sender` helper backs the rate limit.
+- `lambdas/common/friendships_dynamo.py`: added `create_accepted_friendship(sender, recipient)` doing `transact_write_items` on both directional rows with `status='accepted'` and matching `createdAt`/`acceptedAt`.
+- Overwrote handlers: `shares_create`, `shares_feed`, `invites_create`, `invites_accept` (prior shares-and-salvage schema replaced — plan is the Ready source of truth).
+- Created new handlers: `shares_delete/`, `shares_user/` (each with `__init__.py` + `handler.py`).
+- Created 6 test files under `tests/` — 30 tests in total. All pass.
+- Ran full suite: `pytest tests/ --ignore=tests/test_release_radar_dynamo.py --ignore=tests/test_friends_profile.py -q` -> **56 passed**. The two ignored modules fail at import in my local venv for a pre-existing `aiohttp` environment gap unrelated to this work; CI has aiohttp installed.
+- Updated `README.md` with new `/shares/*` and `/invites/*` endpoint sections.
+- Commit `2f7df3b` pushed; PR https://github.com/Xomware/xomify-backend/pull/130 opened. Left unmerged — blocked on PR #67 applying first.
+
+## Decisions / ambiguity resolved
+
+1. **Schema**: followed the plan (denormalized track metadata + moodTag enum + genreTags), overwriting the prior shares-and-salvage schema. Plan is Ready; live DynamoDB table is attribute-agnostic so no storage migration needed.
+2. **403 vs 401 for non-owner delete**: codebase `AuthorizationError` defaults to 401; kept that convention rather than inventing a 403 path. Flagged in PR body.
+3. **`handler=` kwarg**: passed `handler=HANDLER` where the codebase's error classes accept it, but `DynamoDBError` raised inside helpers only gets `function=` and `table=` (matches existing pattern in `friendships_dynamo.py`).
+4. **Pagination `before` semantics**: treated as `ExclusiveStartKey` on the GSI — only items strictly older than `before` come back. `nextBefore` returned when the page is full, otherwise null.
+5. **`invites_accept` 410 vs idempotent 200 on already-friends**: plan's default = 409 Conflict; stuck with that. Open question still tagged in the plan itself.
+6. **`invites_accept` race on concurrent consume**: if the conditional `UpdateItem` fails, returns 410 with `error_code=INVITE_UNAVAILABLE` (distinguishable from the eager 410 paths).
+
+## Results
+
+- Infra PR: https://github.com/Xomware/xomify-infrastructure/pull/67 (1 commit, 1 file, +12 lines)
+- Backend PR: https://github.com/Xomware/xomify-backend/pull/130 (1 commit, 20 files, +660/-210 lines)
+- Tests: 30 new, all passing. Full suite: 56 passing.
+- No blockers hit.
+
+
