+++ docs/features/likes-and-recent-destinations/EXECUTION_LOG.md
@@ -1,0 +1,8 @@
+# Execution Log: Likes & Recently Played — Top-Level Destinations + Friend-Profile Parity
+
+## [2026-04-26 11:41] — Phase 1: Sidebar destinations scaffolded
+
+- **Action**: Added `likes` and `recentlyPlayed` cases to `SidebarDestination`, stub views to `MainShell.destinationRoot`, two drawer entries in `DrawerView`
+- **Files changed**: `NavigationStore.swift`, `MainShell.swift`, `DrawerView.swift`, `Views/Library/LikesView.swift` (new), `Views/Library/RecentlyPlayedView.swift` (new), `project.pbxproj`
+- **Decisions**: Placed Likes + Recently Played adjacent to Feed (after Feed, before Music Taste) per plan default. Stubs return Text placeholders.
+- **Result**: success — PR #85 merged, version bumped to 1.8.0
